basic_server/pyproject.toml
[project]
name = "basic-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.28.0",
    "mcp[cli]>=1.10.1",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.1",
]
